fb_producer/requirements.txt
requests
mysql-connector-python
elasticsearch